django/requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2==2.8.5
psycopg2-binary==2.8.2
dj-database-url==0.5.0